+++ lib/msal-browser/test/interaction_client/NativeInteractionClient.spec.ts
@@ -3,7 +3,7 @@
  * Licensed under the MIT License.
  */
 
-import { AuthenticationScheme, AccountInfo, StubPerformanceClient } from "@azure/msal-common";
+import { AuthenticationScheme, AccountInfo, PromptValue, AuthenticationResult, AccountEntity, IdTokenEntity, AccessTokenEntity, CredentialType, TimeUtils, CacheManager} from "@azure/msal-common";
 import sinon from "sinon";
 import { NativeMessageHandler } from "../../src/broker/nativeBroker/NativeMessageHandler";
 import { ApiId } from "../../src/utils/BrowserConstants";
@@ -11,14 +11,54 @@ import { NativeInteractionClient } from "../../src/interaction_client/NativeInte
 import { PublicClientApplication } from "../../src/app/PublicClientApplication";
 import { ID_TOKEN_CLAIMS, RANDOM_TEST_GUID, TEST_CONFIG, TEST_DATA_CLIENT_INFO, TEST_TOKENS } from "../utils/StringConstants";
 import { NavigationClient } from "../../src/navigation/NavigationClient";
-import { AuthenticationResult } from "@azure/msal-common";
 import { BrowserAuthErrorMessage } from "../../src/error/BrowserAuthError";
-import { PromptValue } from "@azure/msal-common";
 import { NativeAuthError, NativeAuthErrorMessage } from "../../src/error/NativeAuthError";
+import { SilentCacheClient } from "../../src/interaction_client/SilentCacheClient";
+
+const networkInterface = {
+    sendGetRequestAsync<T>(): T {
+        return {} as T;
+    },
+    sendPostRequestAsync<T>(): T {
+        return {} as T;
+    },
+};
+
+const testAccountEntity: AccountEntity = new AccountEntity();
+testAccountEntity.homeAccountId = `${ID_TOKEN_CLAIMS.oid}.${ID_TOKEN_CLAIMS.tid}`;
+testAccountEntity.localAccountId = ID_TOKEN_CLAIMS.oid;
+testAccountEntity.environment = "login.microsoftonline.com";
+testAccountEntity.realm = ID_TOKEN_CLAIMS.tid;
+testAccountEntity.username = ID_TOKEN_CLAIMS.preferred_username;
+testAccountEntity.name = ID_TOKEN_CLAIMS.name;
+testAccountEntity.authorityType = "MSSTS";
+testAccountEntity.nativeAccountId = "nativeAccountId";
+
+const testAccountInfo: AccountInfo = testAccountEntity.getAccountInfo();
+
+const testIdToken: IdTokenEntity = new IdTokenEntity();
+testIdToken.homeAccountId = `${ID_TOKEN_CLAIMS.oid}.${ID_TOKEN_CLAIMS.tid}`;
+testIdToken.clientId = TEST_CONFIG.MSAL_CLIENT_ID;
+testIdToken.environment = testAccountEntity.environment;
+testIdToken.realm = ID_TOKEN_CLAIMS.tid;
+testIdToken.secret = TEST_TOKENS.IDTOKEN_V2;
+testIdToken.credentialType = CredentialType.ID_TOKEN;
+
+const testAccessTokenEntity: AccessTokenEntity = new AccessTokenEntity();
+testAccessTokenEntity.homeAccountId = `${ID_TOKEN_CLAIMS.oid}.${ID_TOKEN_CLAIMS.tid}`;
+testAccessTokenEntity.clientId = TEST_CONFIG.MSAL_CLIENT_ID;
+testAccessTokenEntity.environment = testAccountEntity.environment;
+testAccessTokenEntity.realm = ID_TOKEN_CLAIMS.tid;
+testAccessTokenEntity.secret = TEST_TOKENS.ACCESS_TOKEN;
+testAccessTokenEntity.target = TEST_CONFIG.DEFAULT_SCOPES.join(" ");
+testAccessTokenEntity.credentialType = CredentialType.ACCESS_TOKEN;
+testAccessTokenEntity.expiresOn = `${TimeUtils.nowSeconds() + TEST_CONFIG.TOKEN_EXPIRY}`;
+testAccessTokenEntity.cachedAt = `${TimeUtils.nowSeconds()}`;
+testAccessTokenEntity.tokenType = AuthenticationScheme.BEARER;
 
 describe("NativeInteractionClient Tests", () => {
     globalThis.MessageChannel = require("worker_threads").MessageChannel; // jsdom does not include an implementation for MessageChannel
-    
+
     const pca = new PublicClientApplication({
         auth: {
             clientId: TEST_CONFIG.MSAL_CLIENT_ID
@@ -26,7 +66,7 @@ describe("NativeInteractionClient Tests", () => {
     });
     const wamProvider = new NativeMessageHandler(pca.getLogger(), 2000);
     // @ts-ignore
-    const nativeInteractionClient = new NativeInteractionClient(pca.config, pca.browserStorage, pca.browserCrypto, pca.getLogger(), pca.eventHandler, pca.navigationClient, ApiId.acquireTokenRedirect, pca.performanceClient, wamProvider, "nativeAccountId", RANDOM_TEST_GUID);
+    const nativeInteractionClient = new NativeInteractionClient(pca.config, pca.browserStorage, pca.browserCrypto, pca.getLogger(), pca.eventHandler, pca.navigationClient, ApiId.acquireTokenRedirect, pca.performanceClient, wamProvider, "nativeAccountId", pca.nativeInternalStorage, RANDOM_TEST_GUID);
     let postMessageSpy: sinon.SinonSpy;
     let mcPort: MessagePort;
 
@@ -42,12 +82,47 @@ describe("NativeInteractionClient Tests", () => {
         localStorage.clear();
     });
 
+    describe("acquireTokensFromInternalCache Tests", () => {
+
+        const response: AuthenticationResult = {
+            authority: TEST_CONFIG.validAuthority,
+            uniqueId: testAccountInfo.localAccountId,
+            tenantId: testAccountInfo.tenantId,
+            scopes: TEST_CONFIG.DEFAULT_SCOPES,
+            account: testAccountInfo,
+            idToken: TEST_TOKENS.IDTOKEN_V2,
+            accessToken: TEST_TOKENS.ACCESS_TOKEN,
+            idTokenClaims: ID_TOKEN_CLAIMS,
+            fromCache: true,
+            correlationId: RANDOM_TEST_GUID,
+            expiresOn: new Date(Number(testAccessTokenEntity.expiresOn) * 1000),
+            tokenType: AuthenticationScheme.BEARER
+        };
+
+        sinon.stub(CacheManager.prototype, "readAccountFromCacheWithNativeAccountId").returns(testAccountEntity);
+        sinon.stub(SilentCacheClient.prototype, "acquireToken").callsFake(() => { return Promise.resolve(response); });
+
+        it("Tokens found in cache", async () => {
+            const response = await nativeInteractionClient.acquireToken({ scopes: TEST_CONFIG.DEFAULT_SCOPES });
+            expect(response.accessToken).toEqual(testAccessTokenEntity.secret);
+            expect(response.idToken).toEqual(testIdToken.secret);
+            expect(response.uniqueId).toEqual(ID_TOKEN_CLAIMS.oid);
+            expect(response.tenantId).toEqual(ID_TOKEN_CLAIMS.tid);
+            expect(response.idTokenClaims).toEqual(ID_TOKEN_CLAIMS);
+            expect(response.authority).toEqual(TEST_CONFIG.validAuthority);
+            expect(response.scopes).toEqual(TEST_CONFIG.DEFAULT_SCOPES);
+            expect(response.correlationId).toEqual(RANDOM_TEST_GUID);
+            expect(response.account).toEqual(testAccountEntity.getAccountInfo());
+            expect(response.tokenType).toEqual(AuthenticationScheme.BEARER);
+        });
+    });
+
     describe("acquireToken Tests", () => {
         it("acquires token successfully", async () => {
             const mockWamResponse = {
                 access_token: TEST_TOKENS.ACCESS_TOKEN,
                 id_token: TEST_TOKENS.IDTOKEN_V2,
-                scopes: "User.Read",
+                scope: "User.Read",
                 expires_in: 3600,
                 client_info: TEST_DATA_CLIENT_INFO.TEST_RAW_CLIENT_INFO,
                 account: {
@@ -76,7 +151,7 @@ describe("NativeInteractionClient Tests", () => {
             expect(response.tenantId).toEqual(ID_TOKEN_CLAIMS.tid);
             expect(response.idTokenClaims).toEqual(ID_TOKEN_CLAIMS);
             expect(response.authority).toEqual(TEST_CONFIG.validAuthority);
-            expect(response.scopes).toContain(mockWamResponse.scopes);
+            expect(response.scopes).toContain(mockWamResponse.scope);
             expect(response.correlationId).toEqual(RANDOM_TEST_GUID);
             expect(response.account).toEqual(testAccount);
             expect(response.tokenType).toEqual(AuthenticationScheme.BEARER);
@@ -84,8 +159,8 @@ describe("NativeInteractionClient Tests", () => {
 
         it("throws if prompt: select_account", (done) => {
             nativeInteractionClient.acquireToken({
-                    scopes: ["User.Read"],
-                    prompt: PromptValue.SELECT_ACCOUNT
+                scopes: ["User.Read"],
+                prompt: PromptValue.SELECT_ACCOUNT
             }).catch (e => {
                 expect(e.errorCode).toBe(BrowserAuthErrorMessage.nativePromptNotSupported.code);
                 expect(e.errorMessage).toBe(BrowserAuthErrorMessage.nativePromptNotSupported.desc);
@@ -95,8 +170,8 @@ describe("NativeInteractionClient Tests", () => {
 
         it("throws if prompt: create", (done) => {
             nativeInteractionClient.acquireToken({
-                    scopes: ["User.Read"],
-                    prompt: PromptValue.CREATE
+                scopes: ["User.Read"],
+                prompt: PromptValue.CREATE
             }).catch (e => {
                 expect(e.errorCode).toBe(BrowserAuthErrorMessage.nativePromptNotSupported.code);
                 expect(e.errorMessage).toBe(BrowserAuthErrorMessage.nativePromptNotSupported.desc);
@@ -108,7 +183,7 @@ describe("NativeInteractionClient Tests", () => {
             const mockWamResponse = {
                 access_token: TEST_TOKENS.ACCESS_TOKEN,
                 id_token: TEST_TOKENS.IDTOKEN_V2,
-                scopes: "User.Read",
+                scope: "User.Read",
                 expires_in: 3600,
                 client_info: TEST_DATA_CLIENT_INFO.TEST_RAW_CLIENT_INFO,
                 account: {
@@ -140,7 +215,7 @@ describe("NativeInteractionClient Tests", () => {
             expect(response.tenantId).toEqual(ID_TOKEN_CLAIMS.tid);
             expect(response.idTokenClaims).toEqual(ID_TOKEN_CLAIMS);
             expect(response.authority).toEqual(TEST_CONFIG.validAuthority);
-            expect(response.scopes).toContain(mockWamResponse.scopes);
+            expect(response.scopes).toContain(mockWamResponse.scope);
             expect(response.correlationId).toEqual(RANDOM_TEST_GUID);
             expect(response.account).toEqual(testAccount);
             expect(response.tokenType).toEqual(AuthenticationScheme.BEARER);
@@ -150,7 +225,7 @@ describe("NativeInteractionClient Tests", () => {
             const mockWamResponse = {
                 access_token: TEST_TOKENS.ACCESS_TOKEN,
                 id_token: TEST_TOKENS.IDTOKEN_V2,
-                scopes: "User.Read",
+                scope: "User.Read",
                 expires_in: 3600,
                 client_info: TEST_DATA_CLIENT_INFO.TEST_RAW_CLIENT_INFO,
                 account: {
@@ -182,7 +257,7 @@ describe("NativeInteractionClient Tests", () => {
             expect(response.tenantId).toEqual(ID_TOKEN_CLAIMS.tid);
             expect(response.idTokenClaims).toEqual(ID_TOKEN_CLAIMS);
             expect(response.authority).toEqual(TEST_CONFIG.validAuthority);
-            expect(response.scopes).toContain(mockWamResponse.scopes);
+            expect(response.scopes).toContain(mockWamResponse.scope);
             expect(response.correlationId).toEqual(RANDOM_TEST_GUID);
             expect(response.account).toEqual(testAccount);
             expect(response.tokenType).toEqual(AuthenticationScheme.BEARER);
@@ -192,7 +267,7 @@ describe("NativeInteractionClient Tests", () => {
             const mockWamResponse = {
                 access_token: TEST_TOKENS.ACCESS_TOKEN,
                 id_token: TEST_TOKENS.IDTOKEN_V2,
-                scopes: "User.Read",
+                scope: "User.Read",
                 expires_in: 3600,
                 client_info: TEST_DATA_CLIENT_INFO.TEST_RAW_CLIENT_INFO,
                 account: {
@@ -224,7 +299,7 @@ describe("NativeInteractionClient Tests", () => {
             expect(response.tenantId).toEqual(ID_TOKEN_CLAIMS.tid);
             expect(response.idTokenClaims).toEqual(ID_TOKEN_CLAIMS);
             expect(response.authority).toEqual(TEST_CONFIG.validAuthority);
-            expect(response.scopes).toContain(mockWamResponse.scopes);
+            expect(response.scopes).toContain(mockWamResponse.scope);
             expect(response.correlationId).toEqual(RANDOM_TEST_GUID);
             expect(response.account).toEqual(testAccount);
             expect(response.tokenType).toEqual(AuthenticationScheme.BEARER);
@@ -234,7 +309,7 @@ describe("NativeInteractionClient Tests", () => {
             const mockWamResponse = {
                 access_token: TEST_TOKENS.ACCESS_TOKEN,
                 id_token: TEST_TOKENS.IDTOKEN_V2,
-                scopes: "User.Read",
+                scope: "User.Read",
                 expires_in: 3600,
                 client_info: TEST_DATA_CLIENT_INFO.TEST_RAW_CLIENT_INFO,
                 account: {
@@ -259,7 +334,7 @@ describe("NativeInteractionClient Tests", () => {
             const mockWamResponse = {
                 access_token: TEST_TOKENS.ACCESS_TOKEN,
                 id_token: TEST_TOKENS.IDTOKEN_V2,
-                scopes: "User.Read",
+                scope: "User.Read",
                 expires_in: 3600,
                 client_info: TEST_DATA_CLIENT_INFO.TEST_RAW_CLIENT_INFO,
                 account: {
@@ -283,7 +358,7 @@ describe("NativeInteractionClient Tests", () => {
                 return Promise.resolve(mockWamResponse);
             });
             // @ts-ignore
-            const nativeInteractionClient = new NativeInteractionClient(pca.config, pca.browserStorage, pca.browserCrypto, pca.getLogger(), pca.eventHandler, pca.navigationClient, ApiId.ssoSilent, pca.performanceClient, wamProvider, "nativeAccountId", RANDOM_TEST_GUID);
+            const nativeInteractionClient = new NativeInteractionClient(pca.config, pca.browserStorage, pca.browserCrypto, pca.getLogger(), pca.eventHandler, pca.navigationClient, ApiId.ssoSilent, pca.performanceClient, wamProvider, "nativeAccountId", pca.nativeInternalStorage, RANDOM_TEST_GUID);
             const response = await nativeInteractionClient.acquireToken({
                 scopes: ["User.Read"],
                 prompt: PromptValue.SELECT_ACCOUNT
@@ -294,7 +369,7 @@ describe("NativeInteractionClient Tests", () => {
             expect(response.tenantId).toEqual(ID_TOKEN_CLAIMS.tid);
             expect(response.idTokenClaims).toEqual(ID_TOKEN_CLAIMS);
             expect(response.authority).toEqual(TEST_CONFIG.validAuthority);
-            expect(response.scopes).toContain(mockWamResponse.scopes);
+            expect(response.scopes).toContain(mockWamResponse.scope);
             expect(response.correlationId).toEqual(RANDOM_TEST_GUID);
             expect(response.account).toEqual(testAccount);
             expect(response.tokenType).toEqual(AuthenticationScheme.BEARER);
@@ -304,7 +379,7 @@ describe("NativeInteractionClient Tests", () => {
             const mockWamResponse = {
                 access_token: TEST_TOKENS.ACCESS_TOKEN,
                 id_token: TEST_TOKENS.IDTOKEN_V2,
-                scopes: "User.Read",
+                scope: "User.Read",
                 expires_in: 3600,
                 client_info: TEST_DATA_CLIENT_INFO.TEST_RAW_CLIENT_INFO,
                 account: {
@@ -328,7 +403,7 @@ describe("NativeInteractionClient Tests", () => {
                 return Promise.resolve(mockWamResponse);
             });
             // @ts-ignore
-            const nativeInteractionClient = new NativeInteractionClient(pca.config, pca.browserStorage, pca.browserCrypto, pca.getLogger(), pca.eventHandler, pca.navigationClient, ApiId.acquireTokenSilent_silentFlow, pca.performanceClient, wamProvider, "nativeAccountId", RANDOM_TEST_GUID);
+            const nativeInteractionClient = new NativeInteractionClient(pca.config, pca.browserStorage, pca.browserCrypto, pca.getLogger(), pca.eventHandler, pca.navigationClient, ApiId.acquireTokenSilent_silentFlow, pca.performanceClient, wamProvider, "nativeAccountId", pca.nativeInternalStorage, RANDOM_TEST_GUID);
             const response = await nativeInteractionClient.acquireToken({
                 scopes: ["User.Read"],
                 prompt: PromptValue.SELECT_ACCOUNT
@@ -339,7 +414,7 @@ describe("NativeInteractionClient Tests", () => {
             expect(response.tenantId).toEqual(ID_TOKEN_CLAIMS.tid);
             expect(response.idTokenClaims).toEqual(ID_TOKEN_CLAIMS);
             expect(response.authority).toEqual(TEST_CONFIG.validAuthority);
-            expect(response.scopes).toContain(mockWamResponse.scopes);
+            expect(response.scopes).toContain(mockWamResponse.scope);
             expect(response.correlationId).toEqual(RANDOM_TEST_GUID);
             expect(response.account).toEqual(testAccount);
             expect(response.tokenType).toEqual(AuthenticationScheme.BEARER);
@@ -351,7 +426,7 @@ describe("NativeInteractionClient Tests", () => {
             const mockWamResponse = {
                 access_token: TEST_TOKENS.ACCESS_TOKEN,
                 id_token: TEST_TOKENS.IDTOKEN_V2,
-                scopes: "User.Read",
+                scope: "User.Read",
                 expires_in: 3600,
                 client_info: TEST_DATA_CLIENT_INFO.TEST_RAW_CLIENT_INFO,
                 account: {
@@ -387,7 +462,7 @@ describe("NativeInteractionClient Tests", () => {
             const mockWamResponse = {
                 access_token: TEST_TOKENS.ACCESS_TOKEN,
                 id_token: TEST_TOKENS.IDTOKEN_V2,
-                scopes: "User.Read",
+                scope: "User.Read",
                 expires_in: 3600,
                 client_info: TEST_DATA_CLIENT_INFO.TEST_RAW_CLIENT_INFO,
                 account: {
@@ -424,7 +499,7 @@ describe("NativeInteractionClient Tests", () => {
                 authority: TEST_CONFIG.validAuthority,
                 uniqueId: testAccount.localAccountId,
                 tenantId: testAccount.tenantId,
-                scopes: mockWamResponse.scopes.split(" "),
+                scopes: mockWamResponse.scope.split(" "),
                 idToken: mockWamResponse.id_token,
                 idTokenClaims: ID_TOKEN_CLAIMS,
                 accessToken: mockWamResponse.access_token,
@@ -443,7 +518,7 @@ describe("NativeInteractionClient Tests", () => {
             const mockWamResponse = {
                 access_token: TEST_TOKENS.ACCESS_TOKEN,
                 id_token: TEST_TOKENS.IDTOKEN_V2,
-                scopes: "User.Read",
+                scope: "User.Read",
                 expires_in: 3600,
                 client_info: TEST_DATA_CLIENT_INFO.TEST_RAW_CLIENT_INFO,
                 account: {
@@ -466,7 +541,7 @@ describe("NativeInteractionClient Tests", () => {
             });
             // @ts-ignore
             pca.browserStorage.setInteractionInProgress(true);
-            nativeInteractionClient.acquireTokenRedirect({scopes: ["User.Read"]}).then(() => {
+            nativeInteractionClient.acquireTokenRedirect({ scopes: ["User.Read"] }).then(() => {
                 // @ts-ignore
                 const inProgress = pca.browserStorage.getInteractionInProgress();
                 expect(inProgress).toBeTruthy();
@@ -477,14 +552,14 @@ describe("NativeInteractionClient Tests", () => {
                     expect(isInProgress).toBeFalsy();
                     done();
                 });
-            })
+            });
         });
 
         it("returns null if interaction is not in progress", async () => {
             const mockWamResponse = {
                 access_token: TEST_TOKENS.ACCESS_TOKEN,
                 id_token: TEST_TOKENS.IDTOKEN_V2,
-                scopes: "User.Read",
+                scope: "User.Read",
                 expires_in: 3600,
                 client_info: TEST_DATA_CLIENT_INFO.TEST_RAW_CLIENT_INFO,
                 account: {