+++ lib/msal-common/src/telemetry/performance/PerformanceEvent.ts
@@ -94,6 +94,32 @@ export enum PerformanceEvents {
      * Used to acquire a token from Native component when native brokering is enabled.
      */
     NativeInteractionClientAcquireToken = "nativeInteractionClientAcquireToken",
+
+    /**
+     * Time spent on the network for refresh token acquisition
+     */
+    RefreshTokenClientExecuteTokenRequest = "refreshTokenClientExecuteTokenRequest",
+
+    /**
+     * Time spent creating default headers for requests to token endpoint
+     */
+    BaseClientCreateTokenRequestHeaders = "baseClientCreateTokenRequestHeaders",
+
+    /**
+     * Used to measure the time taken for completing embedded-broker handshake (PW-Broker).
+     */
+    BrokerHandhshake = "brokerHandshake",
+
+    /**
+     * acquireTokenByRefreshToken API in BrokerClientApplication (PW-Broker) .
+     */
+    AcquireTokenByRefreshTokenInBroker = "acquireTokenByRefreshTokenInBroker",
+
+    /**
+     * acquireToken API in BrokerClientApplication.
+     * Used to acquire a token on behalf of the embedded application (PW-Broker).
+     */
+    AcquireTokenByBroker = "acquireTokenByBroker"
 }
 
 /**
@@ -229,6 +255,22 @@ export type PerformanceEvent = {
      * @type {string}
      */
     libraryVersion: string,
+
+    /**
+     * Size of the id token
+     *
+     * @type {number}
+     */
+    idTokenSize?: number,
+
+    /**
+     * 
+     * Size of the access token
+     *
+     * @type {number}
+     */
+
+    accessTokenSize?: number,
 
     /**
      * Application name as specified by the app.